fix: update job instructions for extension-first flow, fix HMDA float parsing
- weekly-scan.md: require extension scrape before pipeline, add
  prerequisite check, remove NY SLA (dataset now requires auth)
- HMDA: change AggregationItem.Sum from int64 to float64 to handle
  scientific notation from CFPB API
- NYSLA: add User-Agent header (still blocked, but try/except in
  phase3 handles it gracefully)

+++ internal/providers/hmda/loans.go
@@ -137,12 +137,12 @@ func makeRunLoansTract(factory ClientFactory) func(*cobra.Command, []string) err
 			if item.CensusTract == tract {
 				avg := int64(0)
 				if item.Count > 0 {
-					avg = item.Sum / int64(item.Count)
+					avg = int64(item.Sum) / int64(item.Count)
 				}
 				summary := TractSummary{
 					Tract:       item.CensusTract,
 					Count:       item.Count,
-					Volume:      item.Sum,
+					Volume:      int64(item.Sum),
 					AvgLoanSize: avg,
 				}
 				return printTractSummary(cmd, summary)
